# openapi.yaml
openapi: 3.0.3
info:
  title: BLPS_Lab3_Monolit API
  description: |
    Описание API для 3 лабораторной работы по БЛПС
    
    **Примечание**: спецификация делалась на скорую руку за пару часов
  version: 1.0.0
servers:
  - url: http://localhost:8081/
components:
  schemas:
...
    Favorite:
      type: object
      properties:
        id:
          type: integer
          example: 1
        shopId:
          type: integer
          example: 1
        shopName:
          type: string
          example: "shop"
    Subscription:
      type: object
      properties:
        id:
          type: integer
          example: 1
        shopId:
          type: integer
          example: 1
        shopName:
          type: string
          example: "shop"
        startDate:
          type: string
          example: "01.01.2024"
        duration:
          type: integer
          example: 100
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
...
paths:
...
  /favorites/by-shop-id/{shopId}:
    get:
      tags:
        - favorites
      summary: Получить избранное пользователя по идентификатору магазина
      operationId: favorite-by-shop-id
      security:
        - basicAuth: []
      parameters:
        - in: path
          name: shopId
          description: shopId
          required: true
          schema:
            type: integer
            format: int64
            example: 1
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Favorite'
        '404':
          description: Favorite not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "error"
  /favorites:
    get:
      tags:
        - favorites
      summary: Просмотр всех избранных магазинов
      operationId: favorites
      security:
        - basicAuth: []
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Favorite'
    post:
      tags:
        - favorites
      summary: Добавление магазина в избранное
      operationId: addToFavorite
      security:
        - basicAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                shopId:
                  type: integer
                  example: 1
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Favorite'
        '404':
          description: Shop not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "error"
  /favorites/{id}:
    delete:
      tags:
        - favorites
      summary: Удаление магазина из избранного
      operationId: removeFromFavorites
      security:
        - basicAuth: []
      parameters:
        - in: path
          name: id
          description: ID
          required: true
          schema:
            type: integer
            format: int64
            example: 1
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Shop removed from favorites"
        '404':
          description: Favorite not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "error"
  /subscriptions/by-shop-id/{shopId}:
    get:
      tags:
        - subscriptions
      summary: Получить подписку пользователя по идентификатору магазина
      operationId: subscription-by-shop-id
      security:
        - basicAuth: []
      parameters:
        - in: path
          name: shopId
          description: shopId
          required: true
          schema:
            type: integer
            format: int64
            example: 1
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subscription'
        '404':
          description: Subscription not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "error"
  /subscriptions:
    get:
      tags:
        - subscriptions
      summary: Просмотр всех активных подписок
      operationId: subscriptions
      security:
        - basicAuth: []
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Subscription'
    post:
      tags:
        - subscriptions
      summary: Оформление/продление подписки на магазин
      operationId: subscribe
      security:
        - basicAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                shopId:
                  type: integer
                  example: 1
                duration:
                  type: integer
                  example: 100
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "message"
        '400':
          description: Incorrect request
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "error"
        '404':
          description: Shop not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "error"
...